chore: Refactor / cleanup.
