+++ data_resource/generator/api_manager/__init__.py
@@ -1,8 +1,3 @@
-from data_resource.generator.api_manager.core import (
-    VersionedResource,
-    VersionedResourceMany,
-)
+from data_resource.generator.api_manager.core import VersionedResource
 from data_resource.generator.api_manager.v1_0_0 import ResourceHandler
-
-
-from data_resource.generator.api_manager.temp import generate_api
+from data_resource.generator.api_manager.api_generator import generate_api
