+++ data_resource/generator/api_manager/core/__init__.py
@@ -1,4 +1,3 @@
 from data_resource.generator.api_manager.core.versioned_resource import (
     VersionedResource,
-    VersionedResourceMany,
 )
