+++ data_resource/generator/api_manager/reference_data_resource_factory.py
