+++ data_resource/generator/api_manager/v1_0_0/crud_functions/__init__.py
@@ -1,0 +1,9 @@
+from data_resource.generator.api_manager.v1_0_0.crud_functions.resource_read import (
+    ResourceRead,
+)
+from data_resource.generator.api_manager.v1_0_0.crud_functions.resource_create import (
+    ResourceCreate,
+)
+from data_resource.generator.api_manager.v1_0_0.crud_functions.resource_update import (
+    ResourceUpdate,
+)
